+++ testsuite/tests/simplCore/should_compile/T12076sat.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE MagicHash #-}
+{-# LANGUAGE BangPatterns, MagicHash #-}
 module T12076sat where
 
 -- This test demonstrates that we need to saturate
@@ -6,4 +6,4 @@ module T12076sat where
 
 import GHC.Exts
 
-f = I# (dataToTag# timesWord#)
+f = I# (case noinline id timesWord# of !_ -> 0#)